build: extract the semantic DOCX/PPTX backend into graph-compose-render-docx
Split the POI-backed semantic office backend out of the engine jar so the lean
core no longer carries Apache POI.

- New standalone render-docx/ module (lockstep-versioned with the engine): moves
  DocxSemanticBackend and the PptxSemanticBackend stub into new format packages
  document.backend.semantic.docx / .pptx (the SPI — SemanticBackend,
  SemanticExportContext, SemanticExportManifest — stays in the core
  document.backend.semantic package; a split package is not allowed). poi-ooxml is
  a plain compile dependency here.
- Root pom drops the <optional> poi-ooxml dependency. The DocumentSession export
  test that exercised both backends moves to render-docx as SessionSemanticExportTest
  (the engine test scope can no longer see the backends); its no-poi guard is
  dropped because POI is always present in this module.
- aggregator adds the module; examples depend on graph-compose-render-docx (which
  brings POI transitively) and WordExportExample imports the new FQN.

Reactor build green: render-docx compiles and its docx tests pass, while the root
builds POI-free.
