+++ aggregator/pom.xml
@@ -60,6 +60,7 @@
         <module>..</module>
         <module>../fonts</module>
         <module>../emoji</module>
+        <module>../render-docx</module>
         <module>../bundle</module>
         <module>../examples</module>
         <module>../benchmarks</module>
