+++ examples/pom.xml
@@ -67,13 +67,13 @@
             <version>${graphcompose.emoji.version}</version>
         </dependency>
 
-        <!-- DOCX export (WordExportExample): poi-ooxml is optional in the
-             GraphCompose POM, so consumers add it explicitly.
-             Keep in lockstep with the root pom's poi.version. -->
+        <!-- DOCX export (WordExportExample): the semantic DOCX backend ships in
+             graph-compose-render-docx since 2.0, which brings Apache POI
+             transitively. -->
         <dependency>
-            <groupId>org.apache.poi</groupId>
-            <artifactId>poi-ooxml</artifactId>
-            <version>5.5.1</version>
+            <groupId>io.github.demchaav</groupId>
+            <artifactId>graph-compose-render-docx</artifactId>
+            <version>${graphcompose.version}</version>
         </dependency>
 
         <dependency>
