+++ pom.xml
@@ -182,19 +182,6 @@
             <version>${flexmark.version}</version>
         </dependency>
 
-        <!--
-            Apache POI powers the canonical DOCX semantic backend. Marked
-            optional so library consumers that only render PDFs do not pull
-            in ~10MB of POI runtime; apps that call DocumentSession.export(new
-            DocxSemanticBackend()) must add poi-ooxml to their own classpath.
-        -->
-        <dependency>
-            <groupId>org.apache.poi</groupId>
-            <artifactId>poi-ooxml</artifactId>
-            <version>${poi.version}</version>
-            <optional>true</optional>
-        </dependency>
-
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-annotations</artifactId>
